% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dean.R
\name{dean}
\alias{dean}
\title{Return apportioned vector according to Dean's Method}
\usage{
dean(x, n = 100, k = 1)
}
\arguments{
\item{x}{A named vector.}

\item{n}{Number of apportioned units to sum to}

\item{k}{The incrementing interval used to search for divisors}
}
\value{
A named vector of length \code{x} containing apportioned integers summing to \code{n}.
}
\description{
Return apportioned vector according to Dean's Method
}
\section{References}{

 http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-deans-method-of-apportionment
}

\section{Further details}{

  Method suggested in 1920s by James Dean, professsor of astronomy and mathematics at Dartmouth and University of Vermont.
}
\examples{
dean(usa1832, 240)
dean(usa1832, 240, 20)
}

