% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hamilton.R
\name{hamilton}
\alias{hamilton}
\title{Return apportioned vector according to Alexander Hamilton's Method}
\usage{
hamilton(x, n = 100)
}
\arguments{
\item{x}{A named vector.}

\item{n}{Number of apportioned units to sum to}
}
\value{
A named vector of length \code{x} containing apportioned integers summing to \code{n}.
}
\description{
Return apportioned vector according to Alexander Hamilton's Method
}
\section{References}{

   http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-hamiltons-method-of-apportionment
}

\section{Further details}{

   Method suggested in 1792 by Alexander Hamilton
}
\examples{
hamilton(usa1990, 435)
}

