% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hhill.R
\name{hhill}
\alias{hhill}
\title{Return apportioned vector according to Huntington-Hill Method}
\usage{
hhill(x, n = 100, k = 1)
}
\arguments{
\item{x}{A named vector.}

\item{n}{Number of apportioned units to sum to}

\item{k}{The incrementing interval used to search for divisors}
}
\value{
A named vector of length \code{x} containing apportioned integers summing to \code{n}.
}
\description{
Return apportioned vector according to Huntington-Hill Method
}
\section{References}{

 http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-hills-method-of-apportionment
}

\section{Further details}{

  Method developed around 1911 by Joseph Hill and modified by Edward Huntington
}
\examples{
hhill(usa1790, 105)
hhill(usa2000, 435)
hhill(usa2000, 435, 20)
}

