% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hhill0.R
\name{hhill0}
\alias{hhill0}
\title{Return apportioned vector according to Huntington-Hill Method - allows for zeros}
\usage{
hhill0(x, n = 100, k = 1)
}
\arguments{
\item{x}{A named vector.}

\item{n}{Number of apportioned units to sum to}

\item{k}{The incrementing interval used to search for divisors}
}
\value{
A named vector of length \code{x} containing apportioned integers summing to \code{n}.
}
\description{
Return apportioned vector according to Huntington-Hill Method - allows for zeros
}
\section{References}{

 http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-hills-method-of-apportionment
}

\section{Further details}{

  Method developed around 1911 by Joseph Hill and modified by Edward Huntington
}
\examples{
majs <- c(100, 154, 0, 22, 5)
names(majs) <- c("English", "History", "Swahili", "French", "Aeronautics")
hhill0(majs,10)
}

