% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jefferson.R
\name{jefferson}
\alias{jefferson}
\title{Return apportioned vector according to Thomas Jefferson's Method}
\usage{
jefferson(x, n = 100, k = 1)
}
\arguments{
\item{x}{A named vector.}

\item{n}{Number of apportioned units to sum to}
}
\value{
A named vector of length \code{x} containing apportioned integers summing to \code{n}.
}
\description{
Return apportioned vector according to Thomas Jefferson's Method
}
\section{References}{

http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-jeffersons-method-of-apportionment
}

\section{Further details}{

  Method suggested in 1792 by Thomas Jefferson
}
\examples{
jefferson(usa1990, 435)
}

