% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jefferson0.R
\name{jefferson0}
\alias{jefferson0}
\title{Return apportioned vector according to Thomas Jefferson's Method - allowing for zeros}
\usage{
jefferson0(x, n = 100, k = 1)
}
\arguments{
\item{x}{A named vector.}

\item{n}{Number of apportioned units to sum to}

\item{k}{The incrementing interval used to search for divisors}
}
\value{
A named vector of length \code{x} containing apportioned integers summing to \code{n}.
}
\description{
Return apportioned vector according to Thomas Jefferson's Method - allowing for zeros
}
\section{References}{

   http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-jeffersons-method-of-apportionment
}

\section{Further details}{

   Method suggested in 1792 by Thomas Jefferson
}
\examples{
z <- c(1,4,5,14,16,3)
names(z) <- letters[1:6]
jefferson0(z, 10, k=0.1)
}

