% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lowndes.R
\name{lowndes}
\alias{lowndes}
\title{Return apportioned vector according to Lowndes' Method}
\usage{
lowndes(x, n = 100)
}
\arguments{
\item{x}{A named vector.}

\item{n}{Number of apportioned units to sum to}
}
\value{
A named vector of length \code{x} containing apportioned integers summing to \code{n}.
}
\description{
Return apportioned vector according to Lowndes' Method
}
\section{References}{

 http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-lowndes-method-of-apportionment
}

\section{Further details}{

  Method suggested in 1822 by Representative William Lowndes of South Carolina
}
\examples{
lowndes(usa1820, 213)
}

