% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parador}
\alias{parador}
\title{Parador example}
\format{A named vector of numbers}
\source{
\url{need_to_relocate}
}
\usage{
parador
}
\description{
Parador example
}
\keyword{datasets}

