% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop1}
\alias{pop1}
\title{Population example 1}
\format{A named vector of numbers}
\source{
\url{need_to_relocate}
}
\usage{
pop1
}
\description{
Population example 1
}
\keyword{datasets}

