% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa1820}
\alias{usa1820}
\title{USA population by state in 1820.}
\format{A named vector of numbers}
\source{
\url{http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-introduction}
}
\usage{
usa1820
}
\description{
USA population by state in 1820.
}
\keyword{datasets}

