% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa2000}
\alias{usa2000}
\title{USA population by state in 2000.}
\format{A named vector of numbers}
\source{
\url{http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-introduction}
}
\usage{
usa2000
}
\description{
USA population by state in 2000.
}
\keyword{datasets}

