% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_census}
\alias{usa_census}
\title{USA population by state in every census.}
\format{A dataframe of states, year admitted to union and population by census year.}
\source{
\url{http://en.wikipedia.org/wiki/List_of_U.S._states_by_historical_population}
}
\usage{
usa_census
}
\description{
USA population by state in every census.
}
\keyword{datasets}

