% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/webster0.R
\name{webster0}
\alias{webster0}
\title{Return apportioned vector according to Webster's Method - allows for zeros}
\usage{
webster0(x, n = 100, k = 1)
}
\arguments{
\item{x}{A named vector.}

\item{n}{Number of apportioned units to sum to}

\item{k}{The incrementing interval used to search for divisors}
}
\value{
A named vector of length \code{x} containing apportioned integers summing to \code{n}.
}
\description{
Return apportioned vector according to Webster's Method - allows for zeros
}
\section{References}{

http://www.maa.org/publications/periodicals/convergence/apportioning-representatives-in-the-united-states-congress-websters-method-of-apportionment
}

\section{Further details}{

 Method suggested by Sen Daniel Webster used in 1840, 1910 and 1930 apportionments
}
\examples{
majs <- c(100, 154, 0, 22, 5)
names(majs) <- c("English", "History", "Swahili", "French", "Aeronautics")
webster0(majs,10)
}

